Move fake transport to test module inside context server
